requirements.txt
requests~=2.27.1
gevent~=21.12.0
beautifulsoup4~=4.10.0
tensorflow-gpu~=2.7.0
django~=4.0.3